// src/react/draw-layer-types.ts
export type DrawCoordinate = [number, number];
// ...
export type DrawOverlayCoordinates = DrawCoordinate[] | DrawCoordinate[][] | DrawCoordinate[][][];

export interface DrawOverlayShape {
  id?: string | number;
  coordinates: DrawOverlayCoordinates;
  closed?: boolean;
  fill?: boolean;
  stroke?: Partial<RegionStrokeStyle>;
  strokeStyle?: Partial<RegionStrokeStyle>;
  invertedFill?: DrawOverlayInvertedFillStyle;
  visible?: boolean;
}

export interface DrawOverlayInvertedFillStyle {
  fillColor: string;
}

export interface RegionStrokeStyle {
  color: string;
  width: number;
  lineDash: number[];
  lineJoin: CanvasLineJoin;
  lineCap: CanvasLineCap;
  shadowColor: string;
  shadowBlur: number;
  shadowOffsetX: number;
  shadowOffsetY: number;
}